+++ lib/drab/config.ex
@@ -2,5 +2,5 @@ defmodule Drab.Config do
   @moduledoc false
 
   # by default load Drab.Query and Drab.Call
-  defstruct commander: nil, onload: nil, modules: [:query, :modal]
+  defstruct commander: nil, onload: nil, modules: [Drab.Query, Drab.Modal]
 end