+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## 0.1.1
+
 Changes:
 * added more jQuery methods, like width, position, etc
 * cycling
@@ -9,3 +10,14 @@ Changes:
 
 Fixes:
 * atom leaking issue (#1)
+
+## 0.2.0
+
+Changes:
+* extracted Drab core; split the JS templates between modules
+* jQuery not required in Drab.Core
+* moved Drab.Query.execjs and broadcastjs to Drab.Core.execjs and broadcastjs
+* moved Drab.Call.console to Drab.Core.console (and console!)
+* renamed Drab.Call to Drab.Modal
+* renamed Drab.Templates to Drab.Template
+* JS: Drab is global, Drab.launch_event() is available